+++ docs/md/releasenotes.md
@@ -10,6 +10,7 @@
 
 - Fix `UnsupportedOperationException` when trying to shrink Kotlin metadata from a lambda function.
 - Change behavior of array index out of bounds during partial evaluation, which is now handled as an unknown value instead of throwing an exception in the partial evaluator method.
+- Fix `IllegalArgumentException` when joining values with different computational type during value analysis.
 
 ### API Changes